requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.0.0
